job #8321 Modified the CutAction's isEnabled() routine to only allow Packages and DataTypes to be moved. Of course Anything is allowed in a package.
